+++ examples/mesh.scm
@@ -1,0 +1,32 @@
+(graphics/mesh 
+    (
+        (
+            (name . "mesh-1") 
+            (surfaces-list 15) 
+            (location-ids) 
+            (locations (physics) (geometry) (surfaces)) 
+            (options 
+                (nodes? . #f) 
+                (edges? . #t) 
+                (faces? . #t) 
+                (partitions? . #f) 
+                (overset? . #f) 
+                (gap? . #f)
+            ) 
+            (edge-type all . #f) 
+            (edge-type-options 
+                (edge-type . "") 
+                (feature-angle . 0.)
+            ) 
+            (shrink-factor . 0.) 
+            (coloring automatic type . #f) 
+            (coloring-options 
+                (option . "") 
+                (faces . "") 
+                (edges . "")
+            ) 
+            (display-state-name . "None") 
+            (annotations-list)
+        )
+    )
+)