+++ examples/scene.scm
@@ -1,0 +1,30 @@
+(graphics/scene 
+    (
+        (
+            (name . "scene-1") 
+            (title . "scene-1") 
+            (temporary? . #f) 
+            (graphics-objects 
+                (
+                    (name . "pathlines-1") 
+                    (transparency . 0) 
+                    (colormap-position . 1) 
+                    (colormap-left . -0.98) 
+                    (colormap-bottom . -0.54) 
+                    (colormap-width . 0.18) 
+                    (colormap-height . 1.08)
+                ) 
+                (
+                    (name . "mesh-1") 
+                    (transparency . 50) 
+                    (colormap-position . -1) 
+                    (colormap-left . 0.) 
+                    (colormap-bottom . 0.) 
+                    (colormap-width . 0.) 
+                    (colormap-height . 0.)
+                )
+            ) 
+            (display-state-name . "None")
+        )
+    )
+)